added create, update and delete post functions
